Reduce the number of Warnings produced (#168)
Make SPL warning free

Restore size_t usage

During a temporary (one hopes) bout of insanity, I had decided that
size_t should be uint32_t and changed all references to it, as well as,
a gringe worthy zsize_t typedef.

This commits reverts all those changes back to upstream `size_t` usage.

One peculiar thing is that the vxproj disables C4242 but it appears to
be "forgotten" halfway through and C4242 Warnings come back. If I use
pragma warning (disable: 4242) individually in each source file it
works. But unsure why the top level project disabling it fails.

+++ ZFSin/spl/module/spl/spl-kstat.c
@@ -1903,8 +1903,8 @@ spl_kstat_fini()
 
 
 void kstat_set_raw_ops(kstat_t *ksp,
-	int(*headers)(char *buf, uint32_t size),
-	int(*data)(char *buf, uint32_t size, void *data),
+	int(*headers)(char *buf, size_t size),
+	int(*data)(char *buf, size_t size, void *data),
 	void *(*addr)(kstat_t *ksp, off_t index))
 {
 }
